Refactor to package with all tests passing

+++ src/kfp_local/run_pipeline.py
@@ -1,21 +1,22 @@
-"""Demonstrating how KFP can be used to work with compied pipelines."""
+"""Demonstrating how KFP can be used to work with compiled pipelines."""
 import json
 import shutil
 import subprocess
+from importlib.resources import files
 from pathlib import Path
 from typing import Any, Protocol
 
-from google.protobuf.json_format import ParseDict, ParseError
+from google.protobuf.json_format import ParseDict
 from kfp.dsl import structures
-from kfp.pipeline_spec.pipeline_spec_pb2 import PipelineSpec
 from kfp.dsl.types.type_utils import (
+    BOOLEAN,
+    LIST,
     NUMBER_DOUBLE,
     NUMBER_INTEGER,
-    BOOLEAN,
     STRING,
-    LIST,
-    STRUCT
+    STRUCT,
 )
+from kfp.pipeline_spec.pipeline_spec_pb2 import PipelineSpec
 
 LOCAL_FOLDER = "object-storage-bucket"
 OUTPUT_METADATA_FILE = "output_metadata.json"
@@ -36,32 +37,34 @@ _ParamType = int | float | str | bool | list | dict
 
 
 def load_pipeline_spec(compiled_pipeline_file: str) -> PipelineSpec:
-    """Load compiled pipeline and parse into PipelineSpec object."""
+    """Load compiled pipeline JSON and parse into PipelineSpec object."""
     pipeline_file = Path.cwd() / compiled_pipeline_file
     if not pipeline_file.exists():
         raise FileNotFoundError(f"Can't find {pipeline_file}")
-    component_dict = structures.load_documents_from_yaml(pipeline_file.read_text())[0]
     try:
+        component_dict = structures.load_documents_from_yaml(pipeline_file.read_text())[0]  # noqa
         pipeline_spec = ParseDict(component_dict, PipelineSpec())
-    except ParseError:
-        raise RuntimeError(f"{pipeline_file} is not a valid `PipelineSpec`")
+    except Exception as e:
+        raise RuntimeError(f"{pipeline_file} is not a valid `PipelineSpec` - {e}")
     return pipeline_spec
 
 
 def get_task_cmd_args(name: str, pipeline: PipelineSpec) -> tuple[list[str], list[str]]:
-    """Return the requested task execution commands."""
+    """Return the container command and args for a stage."""
     executor_name = f"exec-{name}"
     executors = pipeline.deployment_spec.fields["executors"].struct_value.fields
     if executor_name not in executors:
         raise ValueError(f"{name} not found in pipeline")
-    task = executors[f"exec-{name}"].struct_value.fields["container"].struct_value.fields
+    task = (
+        executors[f"exec-{name}"].struct_value.fields["container"].struct_value.fields
+    )
     cmd = [cmd for cmd in task["command"].list_value]
     args = [arg for arg in task["args"].list_value]
     return cmd, args
 
 
 def _extract_value(param_obj: _HasTypeValueAttr, param_type: int) -> _ParamType:
-    """Extract parameer value based on type."""
+    """Extract parameter value based on type."""
     if param_type == NUMBER_INTEGER:
         return int(param_obj.number_value)
     elif param_type == NUMBER_DOUBLE:
@@ -79,8 +82,8 @@ def _extract_value(param_obj: _HasTypeValueAttr, param_type: int) -> _ParamType:
 
 
 def _get_param_value_from_metadata_file(
-        task_name: str, output_key: str = "Output"
-    ) -> _ParamType:
+    task_name: str, output_key: str = "Output"
+) -> _ParamType:
     """Get output parameter from output_metadata.json file."""
     output_metadata_file = Path.cwd() / LOCAL_FOLDER / task_name / OUTPUT_METADATA_FILE
     if not output_metadata_file.exists():
@@ -94,32 +97,29 @@ def _get_param_value_from_metadata_file(
 
 
 def _get_param_value_from_pipeline_inputs(
-        pipeline: PipelineSpec, param_name: str
-    ) -> _ParamType:
+    pipeline: PipelineSpec, param_name: str
+) -> _ParamType:
     """Get pipeline input."""
     pipeline_inputs = pipeline.root.input_definitions.parameters
-    try:
-        param =  pipeline_inputs[param_name]
-    except KeyError:
+    param = pipeline_inputs[param_name]
+    if str(param) == "":
         raise RuntimeError("couldn't find parameter in pipeline inputs")
     return _extract_value(param.default_value, param.parameter_type)
 
 
 def _get_param_value(
-        pipeline: PipelineSpec, task_name: str, param_name: str
-    ) -> _ParamType:
+    pipeline: PipelineSpec, task_name: str, param_name: str
+) -> _ParamType:
     """Find parameter value for a task."""
-    try:
-        component_name = f"comp-{task_name}"
-        component = pipeline.components[component_name]
-        task = pipeline.root.dag.tasks[task_name]
-    except KeyError:
-        RuntimeError(f"{task_name} is not a task in the pipeline specification")
+    component_name = f"comp-{task_name}"
+    component = pipeline.components[component_name]
+    task = pipeline.root.dag.tasks[task_name]
+    if str(component) == "" or str(task) == "":
+        raise RuntimeError(f"{task_name} is not a task in the pipeline specification")
 
-    try:
-        param_type = component.input_definitions.parameters[param_name].parameter_type
-        param = task.inputs.parameters[param_name]
-    except KeyError:
+    param_type = component.input_definitions.parameters[param_name].parameter_type
+    param = task.inputs.parameters[param_name]
+    if str(param_type) == "" or str(param) == "":
         raise RuntimeError(f"Cannot find param={param_name} in task={task_name}")
 
     if str(param.runtime_value):
@@ -127,7 +127,7 @@ def _get_param_value(
     elif str(param.task_output_parameter):
         return _get_param_value_from_metadata_file(
             param.task_output_parameter.producer_task,
-            param.task_output_parameter.output_parameter_key
+            param.task_output_parameter.output_parameter_key,
         )
     elif str(param.component_input_parameter):
         return _get_param_value_from_pipeline_inputs(
@@ -141,6 +141,7 @@ def _get_param_value(
 
 
 def get_func_args(pipeline: PipelineSpec, task_name: str) -> str:
+    """Extract step args from pipeline config."""
     component_name = f"comp-{task_name}"
     component = pipeline.components[component_name]
     input_parameters = [param for param in component.input_definitions.parameters]
@@ -166,36 +167,55 @@ def get_func_args(pipeline: PipelineSpec, task_name: str) -> str:
     executor_args = {
         "inputs": {
             "parameterValues": input_params_spec,
-            "artifacts": input_artifacts_spec
+            "artifacts": input_artifacts_spec,
         },
         "outputs": {
             "artifacts": output_artifacts_spec,
-            "outputFile": f"{LOCAL_FOLDER}/{task_name}/output_metadata.json"
-        }
+            "outputFile": f"{LOCAL_FOLDER}/{task_name}/output_metadata.json",
+        },
     }
     return json.dumps(executor_args)
 
 
 def run_pipeline(
-        compiled_pipeline_json: str = "pipeline.json", use_nox: bool = True
+        dag: list[str], compiled_pipeline: str = "pipeline.json", use_nox: bool = False
     ) -> None:
     """Run a compiled pipeline with default parameter values."""
-    pipeline = load_pipeline_spec(compiled_pipeline_json)
+    pipeline = load_pipeline_spec(compiled_pipeline)
     if pipeline.schema_version != SCHEMA_VERSION:
         msg = (
             f"schema_version={pipeline.schema_version} not supported - please revert to"
             " schema_version={SCHEMA_VERSION} "
         )
         raise RuntimeError(msg)
+    missing_task_defs = [task for task in dag if task not in pipeline.root.dag.tasks]
+    if missing_task_defs:
+        msg = f"missing task defs in pipeline spec: {', '.join(missing_task_defs)}"
+        raise RuntimeError(msg)
     shutil.rmtree(LOCAL_FOLDER, ignore_errors=True)
-    for task in pipeline.root.dag.tasks:
-        cmd, args = get_task_cmd_args(task, pipeline)
-        args[1] = get_func_args(pipeline, task)
-        if use_nox:
-            subprocess.run(["nox", "-s", "run_pipeline_task", "--", *(cmd + args)])
-        else:
-            subprocess.run(cmd + args)
-
+    for task in dag:
+        try:
+            cmd, args = get_task_cmd_args(task, pipeline)
+            args[1] = get_func_args(pipeline, task)
+            if use_nox:
+                noxfile_path = files("kfp_local") / "kfp_noxfile.py"
+                subprocess.run(
+                    ["nox",
+                    "-s",
+                    "run_pipeline_task",
+                    "-f",
+                    noxfile_path,
+                    "--",
+                    *(cmd + args)
+                    ]
+                )
+            else:
+                kfp_module = files("kfp.dsl.types") / "artifact_types.py"
+                subprocess.run(["sed", "-i", r"s/\/gcs\///", kfp_module])
+                breakpoint()
+                subprocess.run(cmd + args)
+        except Exception as e:
+            raise RuntimeError(f"task={task} failed to execute") from e
 
 if __name__ == "__main__":
-    run_pipeline()
+    run_pipeline(["stage-0", "stage-1"], "tests/resources/pipeline.json")
